changed the data type of the PowerOf10 variable

+++ 0x06-pointers_arrays_strings/100-print_number.c
@@ -10,7 +10,7 @@
 void print_number(int Number)
 {
 	int  SolveCharacter;
-	signed long int PowerOf10;
+	signed int PowerOf10;
 	unsigned char State, State0;
 
 	State = PowerOf10 =  State0 = 1, SolveCharacter = 0;
